backend/requirements.txt
absl-py==2.3.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.8.0
eventlet==0.40.3
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.2
fpdf2==2.8.4
fsspec==2025.9.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
polars==1.33.1
protobuf==4.25.8
psutil==7.0.0
pycparser==2.23
pymongo==4.15.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.2
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
requests==2.32.5
pandas==2.2.2
fpdf2==2.8.4